packages/core/pyproject.toml
[tool.poetry]
name = "agent-security-core"
version = "0.1.0"
description = "Core engine for Agent Security Suite"
authors = ["Agent Security Team"]
license = "MIT"
readme = "README.md"
packages = [{include = "agent_core"}]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
pydantic = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
mypy = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
